Ability to map to struct using field name <-> column name
Added a new option to `row.to`, `map`. By default, this is `.ordinal`. When set
to `.name` the column name and field name are used.

There is overhead to using `name` mapping. This can be minimized by creating
a mapper from the result, which internally caches the name -> index lookup.