Stricter type checking when binding.
When binding an int, float or bool, the code no longer defaults to trying to
bind the value. It now always checks for a matching OID parameter type. This may
block some implicit conversions that PG can make, but I'm not sure..and if it
does, those OIDs could be added on a per-case basis.

Helps improve the error messaging raised in:
https://github.com/karlseguin/pg.zig/issues/27