Try to "peek" at query results
PostgreSQL will report some errors when preparing a statement, but other errors
only when executing it. In some cases, this probably makes sense, but in others
it's a bit unexpected.

For example:
select * from does_not_exist <-- fails to prepare

While:
create table already_exists <-- prepares, but fails to execute

We can try to present a more cohensive API by peeking at the first message after
a successful prepare. If it's an error, we can return that as part of the query!
call. If it isn't, we can leave it there for the first call result.next().